Refactors tests in `api` and `authentication` apps. Sets up Django Debug Toolbar
