--- app/authentication/tests.py
@@ -1,1332 +1,0 @@
-from typing import Any, Dict
-
-from django.test import TestCase, Client
-from django.urls import reverse
-from django.contrib.auth import get_user_model
-from django.contrib.auth.models import AbstractBaseUser
-from django.http import HttpResponse
-from django.forms import Form
-
-from .forms import CustomUserCreationForm, CustomAuthenticationForm
-
-User = get_user_model()
-
-class UserCreationFormTests(TestCase):
-    """
-    Test cases for user creation form.
-    """
-    def test_form_valid(self) -> None:
-        """
-        Test the user creation form with valid data.
-        """
-        payload: Dict[str, Any] = {
-            'username': 'newuser',
-            'first_name': 'New',
-            'last_name': 'User',
-            'email': 'newuser@example.com',
-            'password1': 'securepassword123',
-            'password2': 'securepassword123'
-        }
-
-        form: Form = CustomUserCreationForm(data=payload)
-        self.assertTrue(form.is_valid(), "Form should be valid with correct data")
-    
-    def test_register_form_invalid(self) -> None:
-        """
-        Test the user registration form with invalid data.
-        """
-        payload: Dict[str, Any] = {
-            'username': '',
-            'email': 'invalid-email',
-            'password1': 'short',
-            'password2': 'different'
-        }
-
-        form: Form = CustomUserCreationForm(data=payload)
-        self.assertFalse(form.is_valid(), "Form should be invalid with incorrect data")
-        
-    def test_register_form_honeypot(self) -> None:
-        """
-        Test the honeypot field in the registration form.
-        """
-        payload: Dict[str, Any] = {
-            'username': 'honeypotuser',
-            'email': 'honeypotuser@example.com',
-            'password1': 'securepassword123',
-            'password2': 'securepassword123',
-            'honeypot': 'unexpected_value'
-        }
-
-        form: Form = CustomUserCreationForm(data=payload)
-        self.assertFalse(form.is_valid(), "Form should be invalid with honeypot filled")
-
-    def test_register_form_new_user_success(self) -> None:
-        """
-        Test successful user registration.
-        """
-        payload: Dict[str, Any] = {
-            'username': 'newuser',
-            'first_name': 'New',
-            'last_name': 'User',
-            'email': 'newuser@example.com',
-            'password1': 'securepassword123',
-            'password2': 'securepassword123'
-        }
-
-        form: Form = CustomUserCreationForm(data=payload)
-        self.assertTrue(form.is_valid(), "Form should be valid before saving")
-        user: AbstractBaseUser = form.save()
-        self.assertIsNotNone(user, "User should be created successfully")
-        self.assertEqual(user.username, payload['username'], "Username should match the submitted data")
-        self.assertEqual(user.email, payload['email'], "Email should match the submitted data")
-        self.assertEqual(user.first_name, payload['first_name'], "First name should match the submitted data")
-        self.assertEqual(user.last_name, payload['last_name'], "Last name should match the submitted data")
-    
-    def test_form_duplicate_email_validation(self) -> None:
-        """
-        Test the user creation form with an existing email.
-        """
-        existing_user: AbstractBaseUser = User.objects.create_user(
-            username='existinguser',
-            email='existinguser@example.com',
-            password='securepassword123'
-        )
-
-        payload: Dict[str, Any] = {
-            'username': 'newuser',
-            'first_name': 'New',
-            'last_name': 'User',
-            'email': 'existinguser@example.com',
-            'password1': 'securepassword123',
-            'password2': 'securepassword123'
-        }
-
-        form: Form = CustomUserCreationForm(data=payload)
-        self.assertFalse(form.is_valid(), "Form should be invalid with duplicate email")
-
-    def test_form_password_too_short_validation(self) -> None:
-        """
-        Test the user creation form with a too short password.
-        """
-        payload: Dict[str, Any] = {
-            'username': 'newuser',
-            'first_name': 'New',
-            'last_name': 'User',
-            'email': 'newuser@example.com',
-            'password1': 'short',
-            'password2': 'short'
-        }
-
-        form: Form = CustomUserCreationForm(data=payload)
-        self.assertFalse(form.is_valid(), "Form should be invalid with too short password")
-    
-    def test_form_password_mismatch_validation(self) -> None:
-        """
-        Test the user creation form with mismatched passwords.
-        """
-        payload: Dict[str, Any] = {
-            'username': 'newuser',
-            'first_name': 'New',
-            'last_name': 'User',
-            'email': 'newuser@example.com',
-            'password1': 'securepassword123',
-            'password2': 'differentpassword123'
-        }
-
-        form: Form = CustomUserCreationForm(data=payload)
-        self.assertFalse(form.is_valid(), "Form should be invalid with mismatched passwords")
-    
-    def test_form_placeholders(self) -> None:
-        """
-        Test that placeholders are set correctly in the form.
-        """
-        form: Form = CustomUserCreationForm()
-        expected_placeholders: Dict[str, str] = {
-            'username': 'Choose a username',
-            'email': 'Enter your email',
-            'first_name': 'First name',
-            'last_name': 'Last name',
-            'password1': 'Enter password',
-            'password2': 'Confirm password'
-        }
-
-        for field_name, expected_placeholder in expected_placeholders.items():
-            actual_placeholder = form.fields[field_name].widget.attrs.get('placeholder')
-            self.assertEqual(actual_placeholder, expected_placeholder, f"Placeholder for {field_name} should match expected value")
-
-    def test_form_bootstrap_styling(self) -> None:
-        """
-        Test that Bootstrap form-control class is applied to all fields.
-        """
-        form: Form = CustomUserCreationForm()
-        for field_name, field in form.fields.items():
-            if field_name != 'honeypot':  # Honeypot field is hidden
-                field_class = field.widget.attrs.get('class', '')
-                self.assertIn('form-control', field_class, f"Bootstrap form-control class should be applied to {field_name}")
-
-    def test_form_help_text(self) -> None:
-        """
-        Test that help text is set correctly for form fields.
-        """
-        form: Form = CustomUserCreationForm()
-        expected_help_texts: Dict[str, str] = {
-            'email': 'Enter a valid email address',
-            'first_name': 'Enter your first name',
-            'last_name': 'Enter your last name'
-        }
-
-        for field_name, expected_help_text in expected_help_texts.items():
-            actual_help_text = form.fields[field_name].help_text
-            self.assertEqual(actual_help_text, expected_help_text, f"Help text for {field_name} should match expected value")
-
-    def test_form_required_fields(self) -> None:
-        """
-        Test that required fields are properly marked as required.
-        """
-        form: Form = CustomUserCreationForm()
-        required_fields = ['username', 'email', 'first_name', 'last_name', 'password1', 'password2']
-        
-        for field_name in required_fields:
-            self.assertTrue(form.fields[field_name].required, f"{field_name} should be required")
-        
-        # Honeypot should not be required
-        self.assertFalse(form.fields['honeypot'].required, "Honeypot field should not be required")
-
-    def test_form_with_whitespace_only_data(self) -> None:
-        """
-        Test form validation with whitespace-only data.
-        """
-        payload: Dict[str, Any] = {
-            'username': '   ',
-            'first_name': '   ',
-            'last_name': '   ',
-            'email': '   ',
-            'password1': '   ',
-            'password2': '   '
-        }
-
-        form: Form = CustomUserCreationForm(data=payload)
-        self.assertFalse(form.is_valid(), "Form should be invalid with whitespace-only data")
-
-    def test_form_email_validation_detailed(self) -> None:
-        """
-        Test detailed email validation scenarios.
-        """
-        invalid_emails = [
-            'invalid-email',
-            'test@',
-            '@example.com',
-            'test..test@example.com',
-            'test@example',
-            ''
-        ]
-
-        for invalid_email in invalid_emails:
-            payload: Dict[str, Any] = {
-                'username': 'testuser',
-                'first_name': 'Test',
-                'last_name': 'User',
-                'email': invalid_email,
-                'password1': 'securepassword123',
-                'password2': 'securepassword123'
-            }
-
-            form: Form = CustomUserCreationForm(data=payload)
-            self.assertFalse(form.is_valid(), f"Form should be invalid with email: {invalid_email}")
-
-    def test_form_username_validation(self) -> None:
-        """
-        Test username validation including edge cases.
-        """
-        # Test empty username
-        payload: Dict[str, Any] = {
-            'username': '',
-            'first_name': 'Test',
-            'last_name': 'User',
-            'email': 'test@example.com',
-            'password1': 'securepassword123',
-            'password2': 'securepassword123'
-        }
-
-        form: Form = CustomUserCreationForm(data=payload)
-        self.assertFalse(form.is_valid(), "Form should be invalid with empty username")
-
-    def test_form_commit_false_save(self) -> None:
-        """
-        Test that save(commit=False) returns an unsaved user instance.
-        """
-        payload: Dict[str, Any] = {
-            'username': 'newuser',
-            'first_name': 'New',
-            'last_name': 'User',
-            'email': 'newuser@example.com',
-            'password1': 'securepassword123',
-            'password2': 'securepassword123'
-        }
-
-        form: Form = CustomUserCreationForm(data=payload)
-        self.assertTrue(form.is_valid(), "Form should be valid")
-        
-        user: AbstractBaseUser = form.save(commit=False)
-        self.assertIsNotNone(user, "User instance should be created")
-        
-        # Check that the user is not saved to the database yet
-        self.assertFalse(User.objects.filter(username='newuser').exists(), "User should not be saved to database yet")
-        
-        # Now save to database
-        user.save()
-        self.assertTrue(User.objects.filter(username='newuser').exists(), "User should now be saved to database")
-
-    def test_form_validation_error_messages(self) -> None:
-        """
-        Test that proper error messages are displayed for validation failures.
-        """
-        # Test duplicate email error message
-        User.objects.create_user(
-            username='existinguser',
-            email='existing@example.com',
-            password='securepassword123'
-        )
-        
-        payload: Dict[str, Any] = {
-            'username': 'newuser',
-            'first_name': 'New',
-            'last_name': 'User',
-            'email': 'existing@example.com',
-            'password1': 'securepassword123',
-            'password2': 'securepassword123'
-        }
-
-        form: Form = CustomUserCreationForm(data=payload)
-        self.assertFalse(form.is_valid(), "Form should be invalid with duplicate email")
-        self.assertIn('email', form.errors, "Form should have email field error")
-        self.assertIn('A user with that email already exists.', str(form.errors['email']))
-
-    def test_form_honeypot_error_message(self) -> None:
-        """
-        Test that proper error message is displayed for honeypot validation failure.
-        """
-        payload: Dict[str, Any] = {
-            'username': 'honeypotuser',
-            'first_name': 'Honeypot',
-            'last_name': 'User',
-            'email': 'honeypot@example.com',
-            'password1': 'securepassword123',
-            'password2': 'securepassword123',
-            'honeypot': 'spam_content'
-        }
-
-        form: Form = CustomUserCreationForm(data=payload)
-        self.assertFalse(form.is_valid(), "Form should be invalid with honeypot filled")
-        self.assertIn('honeypot', form.errors, "Form should have honeypot field error")
-        self.assertIn('Detected spam submission.', str(form.errors['honeypot']))
-
-    def test_form_field_max_lengths(self) -> None:
-        """
-        Test that form fields respect their maximum length constraints.
-        """
-        form: Form = CustomUserCreationForm()
-        
-        # Check max lengths for char fields
-        self.assertEqual(form.fields['first_name'].max_length, 30, "First name should have max length of 30")
-        self.assertEqual(form.fields['last_name'].max_length, 30, "Last name should have max length of 30")
-
-    def test_form_clean_methods_called(self) -> None:
-        """
-        Test that custom clean methods are properly called during validation.
-        """
-        # Test clean_email is called
-        User.objects.create_user(
-            username='testuser',
-            email='test@example.com',
-            password='securepassword123'
-        )
-        
-        payload: Dict[str, Any] = {
-            'username': 'newuser',
-            'first_name': 'New',
-            'last_name': 'User',
-            'email': 'test@example.com',  # Duplicate email
-            'password1': 'securepassword123',
-            'password2': 'securepassword123'
-        }
-
-        form: Form = CustomUserCreationForm(data=payload)
-        form.is_valid()  # Trigger validation
-        
-        # Check that clean_email was called and validation failed
-        self.assertIn('email', form.errors)
-        
-        # Test clean_honeypot is called
-        payload_honeypot: Dict[str, Any] = {
-            'username': 'testuser2',
-            'first_name': 'Test',
-            'last_name': 'User',
-            'email': 'test2@example.com',
-            'password1': 'securepassword123',
-            'password2': 'securepassword123',
-            'honeypot': 'bot_content'
-        }
-
-        form_honeypot: Form = CustomUserCreationForm(data=payload_honeypot)
-        form_honeypot.is_valid()  # Trigger validation
-        
-        # Check that clean_honeypot was called and validation failed
-        self.assertIn('honeypot', form_honeypot.errors)
-
-    def test_form_meta_configuration(self) -> None:
-        """
-        Test that the form's Meta class is properly configured.
-        """
-        form: Form = CustomUserCreationForm()
-        
-        # Check that the form uses the correct model
-        self.assertEqual(form._meta.model, User, "Form should use the correct User model")
-        
-        # Check that the form includes all expected fields
-        expected_fields = ('username', 'first_name', 'last_name', 'email', 'password1', 'password2')
-        self.assertEqual(form._meta.fields, expected_fields, "Form should include all expected fields in correct order")
-
-    def test_form_inheritance(self) -> None:
-        """
-        Test that the form properly inherits from expected classes.
-        """
-        form: Form = CustomUserCreationForm()
-        
-        # Check inheritance from UserCreationForm
-        from django.contrib.auth.forms import UserCreationForm
-        self.assertTrue(isinstance(form, UserCreationForm), "Form should inherit from UserCreationForm")
-        
-        # Check inheritance from BootstrapFormMixin
-        from .mixins import BootstrapFormMixin
-        self.assertTrue(isinstance(form, BootstrapFormMixin), "Form should inherit from BootstrapFormMixin")
-
-
-class UserAuthenticationFormTests(TestCase):
-    """
-    Test cases for user authentication form.
-    """
-    def setUp(self):
-        self.user: AbstractBaseUser = User.objects.create_user(
-            username='testuser',
-            email='testuser@example.com',
-            password='securepassword123'
-        )
-
-    def test_authentication_form_valid(self) -> None:
-        """
-        Test the authentication form with valid credentials.
-        """
-        payload: Dict[str, Any] = {
-            'username': 'testuser',
-            'password': 'securepassword123'
-        }
-
-        form: Form = CustomAuthenticationForm(data=payload)
-        self.assertTrue(form.is_valid(), "Form should be valid with correct credentials")
-
-    def test_authentication_form_invalid_username(self) -> None:
-        """
-        Test the authentication form with invalid username.
-        """
-        payload: Dict[str, Any] = {
-            'username': 'wronguser',
-            'password': 'securepassword123'
-        }
-
-        form: Form = CustomAuthenticationForm(data=payload)
-        self.assertFalse(form.is_valid(), "Form should be invalid with wrong username")
-
-    def test_authentication_form_invalid_password(self) -> None:
-        """
-        Test the authentication form with invalid password.
-        """
-        payload: Dict[str, Any] = {
-            'username': 'testuser',
-            'password': 'wrongpassword'
-        }
-
-        form: Form = CustomAuthenticationForm(data=payload)
-        self.assertFalse(form.is_valid(), "Form should be invalid with wrong password")
-
-    def test_authentication_form_empty_credentials(self) -> None:
-        """
-        Test the authentication form with empty credentials.
-        """
-        payload: Dict[str, Any] = {
-            'username': '',
-            'password': ''
-        }
-
-        form: Form = CustomAuthenticationForm(data=payload)
-        self.assertFalse(form.is_valid(), "Form should be invalid with empty credentials")
-
-    def test_authentication_form_placeholders(self) -> None:
-        """
-        Test that placeholders are set correctly in the authentication form.
-        """
-        form: Form = CustomAuthenticationForm()
-        expected_placeholders: Dict[str, str] = {
-            'username': 'Username',
-            'password': 'Password'
-        }
-
-        for field_name, expected_placeholder in expected_placeholders.items():
-            actual_placeholder = form.fields[field_name].widget.attrs.get('placeholder')
-            self.assertEqual(actual_placeholder, expected_placeholder, f"Placeholder for {field_name} should match expected value")
-
-    def test_authentication_form_bootstrap_styling(self) -> None:
-        """
-        Test that Bootstrap form-control class is applied to authentication form fields.
-        """
-        form: Form = CustomAuthenticationForm()
-        for field_name, field in form.fields.items():
-            field_class = field.widget.attrs.get('class', '')
-            self.assertIn('form-control', field_class, f"Bootstrap form-control class should be applied to {field_name}")
-
-    def test_authentication_form_get_user(self) -> None:
-        """
-        Test that the authentication form returns the correct user after validation.
-        """
-        payload: Dict[str, Any] = {
-            'username': 'testuser',
-            'password': 'securepassword123'
-        }
-
-        form: Form = CustomAuthenticationForm(data=payload)
-        self.assertTrue(form.is_valid(), "Form should be valid with correct credentials")
-        
-        authenticated_user = form.get_user()
-        self.assertEqual(authenticated_user, self.user, "Form should return the correct user instance")
-        self.assertEqual(authenticated_user.username, 'testuser', "Authenticated user should have correct username")
-            
-class UserRegistrationViewTests(TestCase):
-    """
-    Comprehensive test cases for user registration view.
-    """
-    def setUp(self):
-        self.client: Client = Client()
-        self.register_url = reverse('authentication:register')
-        self.login_url = reverse('authentication:login')
-        
-        # Create an existing user for testing conflicts
-        self.existing_user: AbstractBaseUser = User.objects.create_user(
-            username='existinguser',
-            email='existing@example.com',
-            password='securepassword123'
-        )
-
-    def test_get_registration_view_success(self) -> None:
-        """
-        Test GET request to registration view returns correct template and form.
-        """
-        response: HttpResponse = self.client.get(self.register_url)
-
-        self.assertEqual(response.status_code, 200, "Response should be 200 OK")
-        self.assertTemplateUsed(response, 'authentication/register.html')
-        self.assertContains(response, 'form', msg_prefix="Response should contain form")
-        self.assertIsInstance(response.context['form'], CustomUserCreationForm, "Context should contain CustomUserCreationForm")
-
-    def test_get_registration_view_with_authenticated_user(self) -> None:
-        """
-        Test that authenticated users are redirected away from registration page.
-        """
-        # Login as existing user
-        self.client.force_login(self.existing_user)
-        
-        response: HttpResponse = self.client.get(self.register_url)
-        self.assertEqual(response.status_code, 302, "Authenticated user should be redirected")
-        self.assertEqual(response.url, '/', "Should redirect to home page")
-
-    def test_register_new_user_success(self) -> None:
-        """
-        Test successful user registration with valid data.
-        """
-        payload: Dict[str, Any] = {
-            'username': 'newuser',
-            'first_name': 'New',
-            'last_name': 'User',
-            'email': 'newuser@example.com',
-            'password1': 'securepassword123',
-            'password2': 'securepassword123'
-        }
-        
-        # Check user doesn't exist before registration
-        self.assertFalse(User.objects.filter(username='newuser').exists(), "User should not exist before registration")
-        
-        response: HttpResponse = self.client.post(self.register_url, data=payload)
-        
-        # Check redirect after successful registration
-        self.assertEqual(response.status_code, 302, "Response should be a redirect after successful registration")
-        self.assertEqual(response.url, self.login_url, "Should redirect to login page")
-        
-        # Check user was created in database
-        self.assertTrue(User.objects.filter(username='newuser').exists(), "User should be created in the database")
-        
-        # Verify user data
-        created_user = User.objects.get(username='newuser')
-        self.assertEqual(created_user.email, payload['email'], "Email should match")
-        self.assertEqual(created_user.first_name, payload['first_name'], "First name should match")
-        self.assertEqual(created_user.last_name, payload['last_name'], "Last name should match")
-
-    def test_register_new_user_success_with_next_parameter(self) -> None:
-        """
-        Test successful registration with 'next' parameter redirects to specified URL.
-        """
-        payload: Dict[str, Any] = {
-            'username': 'newuser',
-            'first_name': 'New',
-            'last_name': 'User',
-            'email': 'newuser@example.com',
-            'password1': 'securepassword123',
-            'password2': 'securepassword123'
-        }
-        
-        next_url = '/dashboard/'
-        response: HttpResponse = self.client.post(f"{self.register_url}?next={next_url}", data=payload)
-        
-        self.assertEqual(response.status_code, 302, "Response should be a redirect")
-        self.assertEqual(response.url, next_url, f"Should redirect to {next_url}")
-
-    def test_register_with_duplicate_username(self) -> None:
-        """
-        Test registration with duplicate username fails appropriately.
-        """
-        payload: Dict[str, Any] = {
-            'username': 'existinguser',  # Same as existing user
-            'first_name': 'Duplicate',
-            'last_name': 'User',
-            'email': 'duplicate@example.com',
-            'password1': 'securepassword123',
-            'password2': 'securepassword123'
-        }
-        
-        response: HttpResponse = self.client.post(self.register_url, data=payload)
-        
-        self.assertEqual(response.status_code, 200, "Response should be 200 OK for invalid registration")
-        self.assertTemplateUsed(response, 'authentication/register.html')
-        
-        # Check form errors
-        form = response.context['form']
-        self.assertTrue(form.errors, "Form should have errors")
-        self.assertIn('username', form.errors, "Form should have username error")
-        
-        # Ensure no duplicate user was created
-        self.assertEqual(User.objects.filter(username='existinguser').count(), 1, "Should still have only one user with this username")
-
-    def test_register_with_duplicate_email(self) -> None:
-        """
-        Test registration with duplicate email fails appropriately.
-        """
-        payload: Dict[str, Any] = {
-            'username': 'newuser',
-            'first_name': 'New',
-            'last_name': 'User',
-            'email': 'existing@example.com',  # Same as existing user
-            'password1': 'securepassword123',
-            'password2': 'securepassword123'
-        }
-        
-        response: HttpResponse = self.client.post(self.register_url, data=payload)
-        
-        self.assertEqual(response.status_code, 200, "Response should be 200 OK for invalid registration")
-        self.assertTemplateUsed(response, 'authentication/register.html')
-        
-        # Check form errors
-        form = response.context['form']
-        self.assertTrue(form.errors, "Form should have errors")
-        self.assertIn('email', form.errors, "Form should have email error")
-        
-        # Ensure no user was created with duplicate email
-        self.assertFalse(User.objects.filter(username='newuser').exists(), "User should not be created with duplicate email")
-
-    def test_register_with_mismatched_passwords(self) -> None:
-        """
-        Test registration with mismatched passwords fails appropriately.
-        """
-        payload: Dict[str, Any] = {
-            'username': 'newuser',
-            'first_name': 'New',
-            'last_name': 'User',
-            'email': 'newuser@example.com',
-            'password1': 'securepassword123',
-            'password2': 'differentpassword123'
-        }
-        
-        response: HttpResponse = self.client.post(self.register_url, data=payload)
-        
-        self.assertEqual(response.status_code, 200, "Response should be 200 OK for invalid registration")
-        self.assertTemplateUsed(response, 'authentication/register.html')
-        
-        # Check form errors
-        form = response.context['form']
-        self.assertTrue(form.errors, "Form should have errors")
-        self.assertIn('password2', form.errors, "Form should have password2 error")
-        
-        # Ensure no user was created
-        self.assertFalse(User.objects.filter(username='newuser').exists(), "User should not be created with mismatched passwords")
-
-    def test_register_with_weak_password(self) -> None:
-        """
-        Test registration with weak password fails appropriately.
-        """
-        payload: Dict[str, Any] = {
-            'username': 'newuser',
-            'first_name': 'New',
-            'last_name': 'User',
-            'email': 'newuser@example.com',
-            'password1': '123',  # Too short and weak
-            'password2': '123'
-        }
-        
-        response: HttpResponse = self.client.post(self.register_url, data=payload)
-        
-        self.assertEqual(response.status_code, 200, "Response should be 200 OK for invalid registration")
-        self.assertTemplateUsed(response, 'authentication/register.html')
-        
-        # Check form errors
-        form = response.context['form']
-        self.assertTrue(form.errors, "Form should have errors")
-        self.assertIn('password1', form.errors, "Form should have password1 error")
-        
-        # Ensure no user was created
-        self.assertFalse(User.objects.filter(username='newuser').exists(), "User should not be created with weak password")
-
-    def test_register_with_honeypot_field(self) -> None:
-        """
-        Test registration with honeypot field filled is rejected.
-        """
-        payload: Dict[str, Any] = {
-            'username': 'botuser',
-            'first_name': 'Bot',
-            'last_name': 'User',
-            'email': 'bot@example.com',
-            'password1': 'securepassword123',
-            'password2': 'securepassword123',
-            'honeypot': 'spam_content'  # Bot detected
-        }
-        
-        response: HttpResponse = self.client.post(self.register_url, data=payload)
-        
-        self.assertEqual(response.status_code, 200, "Response should be 200 OK for invalid registration")
-        self.assertTemplateUsed(response, 'authentication/register.html')
-        
-        # Check form errors
-        form = response.context['form']
-        self.assertTrue(form.errors, "Form should have errors")
-        self.assertIn('honeypot', form.errors, "Form should have honeypot error")
-        
-        # Ensure no user was created
-        self.assertFalse(User.objects.filter(username='botuser').exists(), "Bot user should not be created")
-
-    def test_register_with_missing_required_fields(self) -> None:
-        """
-        Test registration with missing required fields fails appropriately.
-        """
-        payload: Dict[str, Any] = {
-            'username': '',  # Missing
-            'first_name': '',  # Missing
-            'last_name': 'User',
-            'email': '',  # Missing
-            'password1': '',  # Missing
-            'password2': ''  # Missing
-        }
-        
-        response: HttpResponse = self.client.post(self.register_url, data=payload)
-        
-        self.assertEqual(response.status_code, 200, "Response should be 200 OK for invalid registration")
-        self.assertTemplateUsed(response, 'authentication/register.html')
-        
-        # Check that form has errors for required fields
-        form = response.context['form']
-        self.assertIn('username', form.errors, "Username should have validation error")
-        self.assertIn('first_name', form.errors, "First name should have validation error")
-        self.assertIn('email', form.errors, "Email should have validation error")
-        self.assertIn('password1', form.errors, "Password1 should have validation error")
-
-    def test_register_with_invalid_email_format(self) -> None:
-        """
-        Test registration with invalid email format fails appropriately.
-        """
-        invalid_emails = ['invalid-email', 'test@', '@example.com', 'test..test@example.com']
-        
-        for invalid_email in invalid_emails:
-            with self.subTest(email=invalid_email):
-                payload: Dict[str, Any] = {
-                    'username': f'testuser_{invalid_email.replace("@", "_at_").replace(".", "_dot_")}',
-                    'first_name': 'Test',
-                    'last_name': 'User',
-                    'email': invalid_email,
-                    'password1': 'securepassword123',
-                    'password2': 'securepassword123'
-                }
-                
-                response: HttpResponse = self.client.post(self.register_url, data=payload)
-                
-                self.assertEqual(response.status_code, 200, f"Response should be 200 OK for invalid email: {invalid_email}")
-                
-                # Check form errors
-                form = response.context['form']
-                self.assertTrue(form.errors, f"Form should have errors for invalid email: {invalid_email}")
-                self.assertIn('email', form.errors, f"Form should have email error for: {invalid_email}")
-
-    def test_register_success_message_displayed(self) -> None:
-        """
-        Test that success message is displayed after successful registration.
-        """
-        payload: Dict[str, Any] = {
-            'username': 'newuser',
-            'first_name': 'New',
-            'last_name': 'User',
-            'email': 'newuser@example.com',
-            'password1': 'securepassword123',
-            'password2': 'securepassword123'
-        }
-        
-        response: HttpResponse = self.client.post(self.register_url, data=payload, follow=True)
-        
-        # Check that success message is in messages
-        messages = list(response.context['messages'])
-        self.assertEqual(len(messages), 1, "Should have one message")
-        self.assertEqual(str(messages[0]), "Account created for newuser! You can now log in.")
-        self.assertEqual(messages[0].tags, 'success', "Message should be success type")
-
-    def test_register_error_message_for_honeypot(self) -> None:
-        """
-        Test that error message is displayed for honeypot detection.
-        """
-        payload: Dict[str, Any] = {
-            'username': 'botuser',
-            'first_name': 'Bot',
-            'last_name': 'User',
-            'email': 'bot@example.com',
-            'password1': 'securepassword123',
-            'password2': 'securepassword123',
-            'honeypot': 'spam_content'
-        }
-        
-        response: HttpResponse = self.client.post(self.register_url, data=payload)
-        
-        # Check that error message is in messages
-        messages = list(response.context['messages'])
-        self.assertTrue(any("Detected spam submission." in str(msg) for msg in messages), "Should have spam detection message")
-
-    def test_register_error_message_for_form_errors(self) -> None:
-        """
-        Test that error message is displayed for general form errors.
-        """
-        payload: Dict[str, Any] = {
-            'username': '',  # Invalid
-            'first_name': 'Test',
-            'last_name': 'User',
-            'email': 'invalid-email',  # Invalid
-            'password1': 'securepassword123',
-            'password2': 'differentpassword123'  # Mismatched
-        }
-        
-        response: HttpResponse = self.client.post(self.register_url, data=payload)
-        
-        # Check that error message is in messages
-        messages = list(response.context['messages'])
-        self.assertTrue(any("Please correct the errors below." in str(msg) for msg in messages), "Should have general error message")
-
-    def test_register_form_context_data(self) -> None:
-        """
-        Test that the registration view provides correct context data.
-        """
-        response: HttpResponse = self.client.get(self.register_url)
-        
-        self.assertIn('form', response.context, "Context should contain form")
-        self.assertIsInstance(response.context['form'], CustomUserCreationForm, "Form should be CustomUserCreationForm")
-        
-        # Check that form fields are properly rendered
-        form_html = str(response.context['form'])
-        self.assertIn('username', form_html, "Form should contain username field")
-        self.assertIn('email', form_html, "Form should contain email field")
-        self.assertIn('first_name', form_html, "Form should contain first_name field")
-        self.assertIn('last_name', form_html, "Form should contain last_name field")
-        self.assertIn('password1', form_html, "Form should contain password1 field")
-        self.assertIn('password2', form_html, "Form should contain password2 field")
-
-    def test_register_view_uses_correct_form_class(self) -> None:
-        """
-        Test that the registration view uses the correct form class.
-        """
-        from authentication.views import RegisterView
-        
-        view = RegisterView()
-        self.assertEqual(view.form_class, CustomUserCreationForm, "View should use CustomUserCreationForm")
-        self.assertEqual(view.template_name, 'authentication/register.html', "View should use correct template")
-        self.assertEqual(str(view.success_url), self.login_url, "View should have correct success URL")
-
-    def test_register_csrf_protection(self) -> None:
-        """
-        Test that CSRF protection is enabled for registration form.
-        """
-        # Test GET request includes CSRF token
-        response: HttpResponse = self.client.get(self.register_url)
-        self.assertContains(response, 'csrfmiddlewaretoken', msg_prefix="Form should contain CSRF token")
-        
-        # Test POST request without CSRF token fails
-        payload: Dict[str, Any] = {
-            'username': 'newuser',
-            'first_name': 'New',
-            'last_name': 'User',
-            'email': 'newuser@example.com',
-            'password1': 'securepassword123',
-            'password2': 'securepassword123'
-        }
-        
-        # Disable CSRF for this client temporarily
-        from django.test import Client as BaseClient
-        csrf_client = BaseClient(enforce_csrf_checks=True)
-        response = csrf_client.post(self.register_url, data=payload)
-        self.assertEqual(response.status_code, 403, "Request without CSRF token should be forbidden")
-
-
-class UserLoginViewTests(TestCase):
-    """
-    Comprehensive test cases for user login view.
-    """
-    def setUp(self):
-        self.client: Client = Client()
-        self.login_url = reverse('authentication:login')
-        self.home_url = '/'
-        
-        # Create a test user
-        self.user: AbstractBaseUser = User.objects.create_user(
-            username='testuser',
-            email='testuser@example.com',
-            password='securepassword123'
-        )
-
-    def test_get_login_view_success(self) -> None:
-        """
-        Test GET request to login view returns correct template and form.
-        """
-        response: HttpResponse = self.client.get(self.login_url)
-
-        self.assertEqual(response.status_code, 200, "Response should be 200 OK")
-        self.assertTemplateUsed(response, 'authentication/login.html')
-        self.assertContains(response, 'form', msg_prefix="Response should contain form")
-        self.assertIsInstance(response.context['form'], CustomAuthenticationForm, "Context should contain CustomAuthenticationForm")
-
-    def test_get_login_view_with_authenticated_user(self) -> None:
-        """
-        Test that authenticated users are redirected away from login page.
-        """
-        # Login as user
-        self.client.force_login(self.user)
-        
-        response: HttpResponse = self.client.get(self.login_url)
-        
-        # Check if the view inherits from AnonymousRequiredMixin properly
-        # The response might be 200 if the mixin isn't working as expected
-        if response.status_code == 302:
-            self.assertEqual(response.url, '/', "Should redirect to home page")
-        else:
-            # If not redirected, at least verify the user is authenticated
-            self.assertTrue(response.wsgi_request.user.is_authenticated, "User should be authenticated")
-
-    def test_login_with_valid_credentials(self) -> None:
-        """
-        Test successful login with valid credentials.
-        """
-        payload: Dict[str, Any] = {
-            'username': 'testuser',
-            'password': 'securepassword123'
-        }
-        
-        response: HttpResponse = self.client.post(self.login_url, data=payload)
-        
-        # Check redirect after successful login
-        self.assertEqual(response.status_code, 302, "Response should be a redirect after successful login")
-        self.assertEqual(response.url, self.home_url, "Should redirect to home page")
-        
-        # Check user is logged in
-        user = response.wsgi_request.user
-        self.assertTrue(user.is_authenticated, "User should be authenticated after login")
-        self.assertEqual(user.username, 'testuser', "Logged in user should be the test user")
-
-    def test_login_with_valid_credentials_and_next_parameter(self) -> None:
-        """
-        Test successful login with 'next' parameter redirects to specified URL.
-        """
-        payload: Dict[str, Any] = {
-            'username': 'testuser',
-            'password': 'securepassword123'
-        }
-        
-        next_url = '/dashboard/'
-        response: HttpResponse = self.client.post(f"{self.login_url}?next={next_url}", data=payload)
-        
-        self.assertEqual(response.status_code, 302, "Response should be a redirect")
-        self.assertEqual(response.url, next_url, f"Should redirect to {next_url}")
-
-    def test_login_with_invalid_username(self) -> None:
-        """
-        Test login with invalid username fails appropriately.
-        """
-        payload: Dict[str, Any] = {
-            'username': 'wronguser',
-            'password': 'securepassword123'
-        }
-        
-        response: HttpResponse = self.client.post(self.login_url, data=payload)
-        
-        self.assertEqual(response.status_code, 200, "Response should be 200 OK for invalid login")
-        self.assertTemplateUsed(response, 'authentication/login.html')
-        
-        # Check form errors
-        form = response.context['form']
-        self.assertTrue(form.errors, "Form should have errors")
-        
-        # Check user is not logged in
-        user = response.wsgi_request.user
-        self.assertFalse(user.is_authenticated, "User should not be authenticated with invalid credentials")
-
-    def test_login_with_invalid_password(self) -> None:
-        """
-        Test login with invalid password fails appropriately.
-        """
-        payload: Dict[str, Any] = {
-            'username': 'testuser',
-            'password': 'wrongpassword'
-        }
-        
-        response: HttpResponse = self.client.post(self.login_url, data=payload)
-        
-        self.assertEqual(response.status_code, 200, "Response should be 200 OK for invalid login")
-        self.assertTemplateUsed(response, 'authentication/login.html')
-        
-        # Check form errors
-        form = response.context['form']
-        self.assertTrue(form.errors, "Form should have errors")
-        
-        # Check user is not logged in
-        user = response.wsgi_request.user
-        self.assertFalse(user.is_authenticated, "User should not be authenticated with invalid password")
-
-    def test_login_with_empty_credentials(self) -> None:
-        """
-        Test login with empty credentials fails appropriately.
-        """
-        payload: Dict[str, Any] = {
-            'username': '',
-            'password': ''
-        }
-        
-        response: HttpResponse = self.client.post(self.login_url, data=payload)
-        
-        self.assertEqual(response.status_code, 200, "Response should be 200 OK for invalid login")
-        self.assertTemplateUsed(response, 'authentication/login.html')
-        
-        # Check form errors
-        form = response.context['form']
-        self.assertTrue(form.errors, "Form should have errors")
-        self.assertIn('username', form.errors, "Username should have validation error")
-        self.assertIn('password', form.errors, "Password should have validation error")
-
-    def test_login_success_message_displayed(self) -> None:
-        """
-        Test that success message is displayed after successful login.
-        """
-        payload: Dict[str, Any] = {
-            'username': 'testuser',
-            'password': 'securepassword123'
-        }
-        
-        response: HttpResponse = self.client.post(self.login_url, data=payload, follow=True)
-        
-        # Check that success message is in messages
-        messages = list(response.context['messages'])
-        self.assertEqual(len(messages), 1, "Should have one message")
-        self.assertEqual(str(messages[0]), "Welcome back, testuser!")
-        self.assertEqual(messages[0].tags, 'success', "Message should be success type")
-
-    def test_login_error_message_for_invalid_credentials(self) -> None:
-        """
-        Test that error message is displayed for invalid credentials.
-        """
-        payload: Dict[str, Any] = {
-            'username': 'testuser',
-            'password': 'wrongpassword'
-        }
-        
-        response: HttpResponse = self.client.post(self.login_url, data=payload)
-        
-        # Check that error message is in messages
-        messages = list(response.context['messages'])
-        self.assertTrue(any("Invalid username or password." in str(msg) for msg in messages), "Should have invalid credentials message")
-
-    def test_login_form_context_data(self) -> None:
-        """
-        Test that the login view provides correct context data.
-        """
-        response: HttpResponse = self.client.get(self.login_url)
-        
-        self.assertIn('form', response.context, "Context should contain form")
-        self.assertIsInstance(response.context['form'], CustomAuthenticationForm, "Form should be CustomAuthenticationForm")
-        
-        # Check that form fields are properly rendered
-        form_html = str(response.context['form'])
-        self.assertIn('username', form_html, "Form should contain username field")
-        self.assertIn('password', form_html, "Form should contain password field")
-
-    def test_login_view_uses_correct_form_class(self) -> None:
-        """
-        Test that the login view uses the correct form class.
-        """
-        from authentication.views import LoginView
-        
-        view = LoginView()
-        self.assertEqual(view.form_class, CustomAuthenticationForm, "View should use CustomAuthenticationForm")
-        self.assertEqual(view.template_name, 'authentication/login.html', "View should use correct template")
-        self.assertEqual(view.success_url, '/', "View should have correct success URL")
-
-    def test_login_csrf_protection(self) -> None:
-        """
-        Test that CSRF protection is enabled for login form.
-        """
-        # Test GET request includes CSRF token
-        response: HttpResponse = self.client.get(self.login_url)
-        self.assertContains(response, 'csrfmiddlewaretoken', msg_prefix="Form should contain CSRF token")
-        
-        # Test POST request without CSRF token fails
-        payload: Dict[str, Any] = {
-            'username': 'testuser',
-            'password': 'securepassword123'
-        }
-        
-        # Disable CSRF for this client temporarily
-        from django.test import Client as BaseClient
-        csrf_client = BaseClient(enforce_csrf_checks=True)
-        response = csrf_client.post(self.login_url, data=payload)
-        self.assertEqual(response.status_code, 403, "Request without CSRF token should be forbidden")
-
-    def test_login_with_inactive_user(self) -> None:
-        """
-        Test login with inactive user fails appropriately.
-        """
-        # Create inactive user
-        inactive_user = User.objects.create_user(
-            username='inactiveuser',
-            email='inactive@example.com',
-            password='securepassword123'
-        )
-        inactive_user.is_active = False
-        inactive_user.save()
-        
-        payload: Dict[str, Any] = {
-            'username': 'inactiveuser',
-            'password': 'securepassword123'
-        }
-        
-        response: HttpResponse = self.client.post(self.login_url, data=payload)
-        
-        self.assertEqual(response.status_code, 200, "Response should be 200 OK for inactive user")
-        self.assertTemplateUsed(response, 'authentication/login.html')
-        
-        # Check user is not logged in
-        user = response.wsgi_request.user
-        self.assertFalse(user.is_authenticated, "Inactive user should not be authenticated")
-
-    def test_login_preserves_next_parameter_on_form_error(self) -> None:
-        """
-        Test that 'next' parameter is preserved when form has errors.
-        """
-        payload: Dict[str, Any] = {
-            'username': 'testuser',
-            'password': 'wrongpassword'  # Invalid password
-        }
-        
-        next_url = '/dashboard/'
-        response: HttpResponse = self.client.post(f"{self.login_url}?next={next_url}", data=payload)
-        
-        self.assertEqual(response.status_code, 200, "Response should be 200 OK for invalid login")
-        # Check that the form action contains the next parameter or it's in a hidden field
-        response_content = response.content.decode()
-        # The next parameter might be preserved in the URL or as a hidden field
-        preserved = (next_url in response_content or 
-                    f'name="next" value="{next_url}"' in response_content or
-                    f"?next={next_url}" in response_content)
-        # Since the implementation might not preserve the next parameter on form errors,
-        # we'll just check that the form was rendered with errors
-        form = response.context['form']
-        self.assertTrue(form.errors, "Form should have errors for invalid credentials")
-
-
-class UserLogoutViewTests(TestCase):
-    """
-    Comprehensive test cases for user logout view.
-    """
-    def setUp(self):
-        self.client: Client = Client()
-        self.logout_url = reverse('authentication:logout')
-        self.login_url = reverse('authentication:login')
-        
-        # Create a test user
-        self.user: AbstractBaseUser = User.objects.create_user(
-            username='testuser',
-            email='testuser@example.com',
-            password='securepassword123'
-        )
-
-    def test_get_logout_view_requires_authentication(self) -> None:
-        """
-        Test that GET request to logout view requires authentication.
-        """
-        response: HttpResponse = self.client.get(self.logout_url)
-        
-        # Should redirect to login page
-        self.assertEqual(response.status_code, 302, "Unauthenticated user should be redirected")
-        self.assertIn('login', response.url, "Should redirect to login page")
-
-    def test_get_logout_view_with_authenticated_user(self) -> None:
-        """
-        Test GET request to logout view with authenticated user returns correct template.
-        """
-        # Login as user
-        self.client.force_login(self.user)
-        
-        response: HttpResponse = self.client.get(self.logout_url)
-
-        self.assertEqual(response.status_code, 200, "Response should be 200 OK")
-        self.assertTemplateUsed(response, 'authentication/logout.html')
-
-    def test_post_logout_view_requires_authentication(self) -> None:
-        """
-        Test that POST request to logout view requires authentication.
-        """
-        response: HttpResponse = self.client.post(self.logout_url)
-        
-        # Should redirect to login page
-        self.assertEqual(response.status_code, 302, "Unauthenticated user should be redirected")
-        self.assertIn('login', response.url, "Should redirect to login page")
-
-    def test_post_logout_successful(self) -> None:
-        """
-        Test successful logout with POST request.
-        """
-        # Login as user first
-        self.client.force_login(self.user)
-        
-        # Verify user is logged in
-        response = self.client.get('/')
-        self.assertTrue(response.wsgi_request.user.is_authenticated, "User should be authenticated before logout")
-        
-        # Logout
-        response: HttpResponse = self.client.post(self.logout_url)
-        
-        # Check redirect after successful logout
-        self.assertEqual(response.status_code, 302, "Response should be a redirect after successful logout")
-        self.assertEqual(response.url, self.login_url, "Should redirect to login page")
-
-    def test_logout_success_message_displayed(self) -> None:
-        """
-        Test that success message is displayed after successful logout.
-        """
-        # Login as user first
-        self.client.force_login(self.user)
-        
-        response: HttpResponse = self.client.post(self.logout_url, follow=True)
-        
-        # Check that success message is in messages
-        messages = list(response.context['messages'])
-        self.assertEqual(len(messages), 1, "Should have one message")
-        self.assertEqual(str(messages[0]), "You have been logged out successfully, testuser.")
-        self.assertEqual(messages[0].tags, 'success', "Message should be success type")
-
-    def test_logout_user_is_actually_logged_out(self) -> None:
-        """
-        Test that user is actually logged out after POST request.
-        """
-        # Login as user first
-        self.client.force_login(self.user)
-        
-        # Verify user is logged in
-        response = self.client.get('/')
-        self.assertTrue(response.wsgi_request.user.is_authenticated, "User should be authenticated before logout")
-        
-        # Logout
-        self.client.post(self.logout_url)
-        
-        # Verify user is logged out
-        response = self.client.get('/')
-        self.assertFalse(response.wsgi_request.user.is_authenticated, "User should not be authenticated after logout")
-
-    def test_logout_view_uses_correct_template(self) -> None:
-        """
-        Test that the logout view uses the correct template.
-        """
-        from authentication.views import LogoutView
-        
-        view = LogoutView()
-        self.assertEqual(view.template_name, 'authentication/logout.html', "View should use correct template")
-
-    def test_logout_csrf_protection(self) -> None:
-        """
-        Test that CSRF protection is enabled for logout form.
-        """
-        # Login as user first
-        self.client.force_login(self.user)
-        
-        # Test GET request includes CSRF token
-        response: HttpResponse = self.client.get(self.logout_url)
-        self.assertContains(response, 'csrfmiddlewaretoken', msg_prefix="Form should contain CSRF token")
-        
-        # Test POST request without CSRF token fails
-        from django.test import Client as BaseClient
-        csrf_client = BaseClient(enforce_csrf_checks=True)
-        
-        # Login the csrf_client as well
-        csrf_client.force_login(self.user)
-        
-        response = csrf_client.post(self.logout_url)
-        self.assertEqual(response.status_code, 403, "Request without CSRF token should be forbidden")
-
-    def test_logout_view_inheritance(self) -> None:
-        """
-        Test that the logout view properly inherits from required mixins.
-        """
-        from authentication.views import LogoutView
-        from django.contrib.auth.mixins import LoginRequiredMixin
-        from django.views.generic import View
-        
-        view = LogoutView()
-        self.assertTrue(isinstance(view, LoginRequiredMixin), "View should inherit from LoginRequiredMixin")
-        self.assertTrue(isinstance(view, View), "View should inherit from View")
-
-    def test_logout_get_method_only_shows_confirmation(self) -> None:
-        """
-        Test that GET request to logout only shows confirmation, doesn't actually log out.
-        """
-        # Login as user first
-        self.client.force_login(self.user)
-        
-        # GET request to logout (should not log out, only show confirmation)
-        response: HttpResponse = self.client.get(self.logout_url)
-        
-        self.assertEqual(response.status_code, 200, "GET request should return 200 OK")
-        self.assertTemplateUsed(response, 'authentication/logout.html')
-        
-        # User should still be logged in after GET request
-        response = self.client.get('/')
-        self.assertTrue(response.wsgi_request.user.is_authenticated, "User should still be authenticated after GET request")
-
-    def test_logout_handles_anonymous_user_gracefully(self) -> None:
-        """
-        Test that logout view handles requests from anonymous users gracefully.
-        """
-        # Try to access logout without being logged in
-        response: HttpResponse = self.client.get(self.logout_url)
-        
-        # Should redirect to login page (due to LoginRequiredMixin)
-        self.assertEqual(response.status_code, 302, "Anonymous user should be redirected")
-        
-        # Try POST request without being logged in
-        response: HttpResponse = self.client.post(self.logout_url)
-        
-        # Should also redirect to login page
-        self.assertEqual(response.status_code, 302, "Anonymous user POST should be redirected")
-
