+++ app/config/urls.py
@@ -1,21 +1,10 @@
-"""
-URL configuration for config project.
-
-The `urlpatterns` list routes URLs to views. For more information please see:
-    https://docs.djangoproject.com/en/5.2/topics/http/urls/
-Examples:
-Function views
-    1. Add an import:  from my_app import views
-    2. Add a URL to urlpatterns:  path('', views.home, name='home')
-Class-based views
-    1. Add an import:  from other_app.views import Home
-    2. Add a URL to urlpatterns:  path('', Home.as_view(), name='home')
-Including another URLconf
-    1. Import the include() function: from django.urls import include, path
-    2. Add a URL to urlpatterns:  path('blog/', include('blog.urls'))
-"""
 from django.contrib import admin
 from django.urls import path, include
+from django.conf import settings
+from dotenv import load_dotenv
+import os
+
+load_dotenv()
 
 urlpatterns = [
     path("admin/", admin.site.urls),
@@ -23,3 +12,14 @@ urlpatterns = [
     path("api/", include("api.urls")),
     path("", include("journal.urls")),
 ]
+
+# Debug toolbar URLs - check both settings.DEBUG and environment variable
+if settings.DEBUG or (hasattr(settings, 'DEBUG') and settings.DEBUG) or os.getenv("DEBUG") == "True":
+    try:
+        import debug_toolbar
+        urlpatterns = [
+            path("__debug__/", include(debug_toolbar.urls)),
+        ] + urlpatterns
+    except ImportError:
+        # Debug toolbar not installed, skip
+        pass
