+++ docker-compose.yml
@@ -18,6 +18,8 @@ services:
       python manage.py runserver 0.0.0.0:8000"
     env_file:
       - .env
+    extra_hosts:
+      - "host.docker.internal:host-gateway"
 
   db:
     image: postgres:17
