+++ requirements.txt
@@ -3,3 +3,5 @@ djangorestframework==3.16.0
 pyeditorjs==1.0.0b0
 python-dotenv==1.1.1
 psycopg2==2.9.10
+drf-spectacular==0.28.0
+django-debug-toolbar==6.0.0
